Moved dataloader functon to a new utils.py module that can be shared among the different subpackages
