+++ PINNs/pinns/lebb/utils.py
@@ -1,32 +1,12 @@
 from collections.abc import Callable
 from typing import Tuple
 
-import numpy as np
-
 import jax
 import jax.numpy as jnp
 from jaxtyping import Array
 
 from .config import Config
 from .cases import get_bc_case
-
-# === -------------------------------------------------------------------- === #
-# dataloader
-# === -------------------------------------------------------------------- === #
-
-def dataloader(arrays, batch_size):
-    dataset_size = arrays[0].shape[0]
-    assert all(array.shape[0] == dataset_size for array in arrays)
-    indices = np.arange(dataset_size)
-    while True:
-        perm = np.random.permutation(indices)
-        start = 0
-        end = batch_size
-        while end <= dataset_size:
-            batch_perm = perm[start:end]
-            yield tuple(array[batch_perm] for array in arrays)
-            start = end
-            end = start + batch_size
 
 
 # === -------------------------------------------------------------------- === #
