+++ PINNs/pinns/utils.py
@@ -1,0 +1,19 @@
+import numpy as np
+
+# === -------------------------------------------------------------------- === #
+# dataloader
+# === -------------------------------------------------------------------- === #
+
+def dataloader(arrays, batch_size):
+    dataset_size = arrays[0].shape[0]
+    assert all(array.shape[0] == dataset_size for array in arrays)
+    indices = np.arange(dataset_size)
+    while True:
+        perm = np.random.permutation(indices)
+        start = 0
+        end = batch_size
+        while end <= dataset_size:
+            batch_perm = perm[start:end]
+            yield tuple(array[batch_perm] for array in arrays)
+            start = end
+            end = start + batch_size
